Split main view code from main.lua into runView.lua.
